move datareader to its own mod
